MLMediaLibraryKit: fix duplication of Files
When Files with a cyrillic characters were used, the relativepath and path component
would hold differently compositions(codepoints) of that character, which led to a mismatch when we
tried to query for the files with this path. Using the decomposed canonical equivalent will return a normalized String
after Unicode Normalization Form D which ensures that we have an identical String no matter the composition.